Create README.md
Tune-hub stands as a vibrant and dynamic Full Stack Web Application, skillfully crafted with the robust combination of Spring Boot and Hibernate. This music streaming platform not only provides a seamless musical journey for users but also strategically integrates the Razer Pay API, elevating the experience for both Agents and Customers.